requirements.txt
torch
torchvision
matplotlib
pandas
Pillow
scikit-learn
shap
